[C++ API] Expose AnyValue and AnyModuleHolder classes
